Changed CXO pics
This patch crops the current pics so that Simon's pic is better centred.
Consequentially, I also had to crop Daniel's pic so that they match in size and
proportions. This results in slightly smaller pictures also on the web page,
but that is also an improvement IMO.

Additionally, I applied some very light cosmetics to both images, to mask skin
irregularities and make the eyes and overall facial appearance a little
crisper.

Closes #17