Added text beratung.md
This patch introduces a new text for the page 'Beratung'.

It adds an introducotry paragraph to the before existing text of the worklflow description.

I still would very much like to add a simple flow chart of the worklfow to accompany the text, in place of the numbered listing that is now there.